Added required polyfills for ie11 added remote plugin code to container
