+++ package.json
@@ -3,10 +3,6 @@
   "version": "2.0.0",
   "description": "The iframe controller for interacting with SpringRoll applications",
   "main": "./dist/index.js",
-  "author": {
-    "name": "Matt Karl",
-    "email": "matt@cloudkid.com"
-  },
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -15,7 +11,6 @@
   "devDependencies": {
     "@babel/core": "^7.1.0",
     "@babel/plugin-external-helpers": "^7.0.0",
-    "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
     "babel-loader": "^8.0.2",
     "chai": "^4.1.2",
@@ -34,7 +29,6 @@
     "karma-firefox-launcher": "^1.1.0",
     "karma-html": "^1.0.5",
     "karma-mocha": "^1.3.0",
-    "karma-polyfill": "^1.0.0",
     "karma-requirejs": "^1.1.0",
     "karma-webpack": "^4.0.0-rc.2",
     "mocha": "^5.2.0",
@@ -50,8 +44,13 @@
     "rollup-plugin-terser": "^3.0.0"
   },
   "dependencies": {
+    "array-foreach": "^1.0.2",
     "bellhop-iframe": "^2.2.0",
-    "nodelist-foreach-polyfill": "^1.2.0"
+    "custom-event-polyfill": "^1.0.6",
+    "es6-object-assign": "^1.1.0",
+    "nodelist-foreach-polyfill": "^1.2.0",
+    "promise-polyfill": "^8.1.0",
+    "whatwg-fetch": "^3.0.0"
   },
   "keywords": [
     "deployment",
